requirements.txt
streamlit==1.29.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
scikit-learn==1.3.0
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
